final commit

+++ backend/db/repository/blog.py
@@ -22,20 +22,42 @@ def list_blogs(db: Session):
     return blogs
 #note also here we don't have auto update for equal id
 #have to retrieve
-def update_blog(id:int, blog: UpdateBlog, author_id: int, db: Session):
+# def update_blog(id:int, blog: UpdateBlog, author_id: int, db: Session):
+#     blog_in_db = db.query(Blog).filter(Blog.id == id).first()
+#     if not blog_in_db:
+#         return
+#     blog_in_db.title = blog.title
+#     blog_in_db.content = blog.content
+#     db.add(blog_in_db)
+#     db.commit()
+#     return blog_in_db
+#
+# def delete_blog(id:int, author_id:int,db:Session):
+#     blog_in_db = db.query(Blog).filter(Blog.id == id)
+#     if not blog_in_db.first():
+#         return {"error":f"Could not find blog with id {id}"}
+#     blog_in_db.delete()
+#     db.commit()
+#     return {"msg":f"deleted blog with id {id}"}
+def update_blog(id: int, blog: UpdateBlog, author_id: int, db: Session):
     blog_in_db = db.query(Blog).filter(Blog.id == id).first()
     if not blog_in_db:
-        return
+        return {"error":f"Blog with id {id} does not exist"}
+    if not blog_in_db.author_id == author_id:                   #new
+        return {"error":f"Only the author can modify the blog"}
     blog_in_db.title = blog.title
     blog_in_db.content = blog.content
     db.add(blog_in_db)
     db.commit()
     return blog_in_db
 
-def delete_blog(id:int, author_id:int,db:Session):
+
+def delete_blog(id: int, author_id: int, db: Session):
     blog_in_db = db.query(Blog).filter(Blog.id == id)
     if not blog_in_db.first():
-        return {"error":f"Could not find blog with id {id}"}
+        return {"error": f"Could not find blog with id {id}"}
+    if not blog_in_db.first().author_id ==author_id:             #new
+        return {"error":f"Only the author can delete a blog"}
     blog_in_db.delete()
     db.commit()
-    return {"msg":f"deleted blog with id {id}"}
+    return {"msg": f"deleted blog with id {id}"}
